Refactor python
